+++ tests/cli/test_nodes_cli.py
@@ -5,19 +5,20 @@ from lavaclient.api.response import Node
 
 
 @patch('sys.argv', ['lava', 'nodes', 'list', 'cluster_id'])
-@patch('lavaclient.api.response.print_table')
-def test_list(resp_print_table, print_table, mock_client, nodes_response):
+def test_list(print_table, mock_client, nodes_response):
     mock_client._request.return_value = nodes_response
     main()
 
-    (data, header), kwargs = print_table.call_args
+    assert print_table.call_count == 2
+
+    (data, header), kwargs = print_table.call_args_list[0]
     alldata = [entry for entry in list(data)[0]]
     assert alldata[:6] == ['node_id', 'NODENAME', 'node_group', 'ACTIVE',
                            '1.2.3.4', '5.6.7.8']
     assert header == Node.table_header
     assert kwargs['title'] == 'Nodes'
 
-    (data, header), kwargs = resp_print_table.call_args
+    (data, header), kwargs = print_table.call_args_list[1]
     alldata = [entry for entry in list(data)[0]]
     assert alldata[:6] == ['NODENAME', 'component_name', 'Component name',
                            'http://host']