+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 -----
 * CLI
     * Fixed --version flag
+    * Updated cluster node groups format to show component names instead of
+      raw JSON
 * You may now SSH to any cluster not in ERROR status, including IMPAIRED
   clusters